requirements.txt
fastapi==0.100.0
uvicorn[standard]==0.23.1
motor==3.6.0
pydantic==1.10.11
python-dotenv==1.0.0
bcrypt==4.2.0
PyJWT==2.9.0
Pika==1.3.2
python-socketio[asgi]==5.7.2